Redirect users to /login instead of /manage
The latter resulted in a message being displayed indicating
they had to be authenticated to access it. Duh, we just
logged out. Instead, we simply redirect the user to /login
and flash a message stating they've been successfully logged
out (if they actually had an identity, and not an
instance of `AnonymousUserMixin`.