Check for and reload rules when we get a log line
Essentially, when the management frontend updates a rule, we need to
ensure that our background tasks will reload it properly.  In a perfect
world, what we'd do is have the frontend simply notify the backend that
it needs to reload, which it then does.

Historically, that's what we've done: we had a `reload_rules` task that
would force the manager to load rules from the database.  However, this
doesn't work for a combination of 2 reasons:
1. A Celery task, by default, goes to a single worker
2. Broadcast tasks do not work when using the Redis broker in Celery
   (see https://github.com/celery/celery/issues/2638)

As such, there's no good way to have the Flask frontend notify all
workers that they need to update their rules.

The solution is to have the worker(s) reload all rules every time they
get a new log line.  This does put no small amount of load on the DB,
but we implement a somewhat stupid caching mechanism that at least
prevents us from re-fetching all the rules every time.

As a long-term solution, we should figure out a way to either get
broadcast tasks working in Celery, or use another notification method.